requirements.txt
ophyd>=1.6.0
bluesky>=1.8.0
numpy>=1.19.0
matplotlib>=3.3.0
pytest
pyepics

neutron-tavi>=0.3.0
